Do not convert DateTimeField and DateField field into DB values
Psycopg2 can handle it itself. It fixes a bug with
insert_or_update_many where it was failing to find existing rows
if a datetime field was present in the update keys. This type of
field was returned as a datetime.datetime object from psycopg2
select query, while the code was casting the field to a DB value,
(plain text).